add function booking

+++ src/main/java/qnt/moviebooking/dto/request/ShowtimeRequestDto.java
@@ -7,6 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -17,7 +18,7 @@ import java.util.List;
 public class ShowtimeRequestDto {
     private Long movieId;
     private Long auditoriumId;
-    private Integer basePrice;
+    private BigDecimal basePrice;
     @JsonFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     private List<LocalDateTime> startTimes;
 }
